backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
django-cors-headers==4.1.0
djoser==2.2.0
Pillow==10.0.0
PyJWT==2.7.0 